refactor: abstract client-side proof logic
