+++ x/sync/db.go
@@ -3,7 +3,14 @@
 
 package sync
 
-import "github.com/ava-labs/avalanchego/x/merkledb"
+import (
+	"context"
+
+	"github.com/ava-labs/avalanchego/utils/maybe"
+	"github.com/ava-labs/avalanchego/x/merkledb"
+
+	pb "github.com/ava-labs/avalanchego/proto/pb/sync"
+)
 
 type DB interface {
 	merkledb.Clearer
@@ -11,4 +18,17 @@ type DB interface {
 	merkledb.ProofGetter
 	merkledb.ChangeProofer
 	merkledb.RangeProofer
+}
+
+type ProofClient interface {
+	merkledb.Clearer
+	merkledb.MerkleRootGetter
+	HandleRangeProofResponse(ctx context.Context, request *pb.SyncGetRangeProofRequest, responseBytes []byte, onFinish func(maybe.Maybe[[]byte])) error
+	HandleChangeProofResponse(
+		ctx context.Context,
+		request *pb.SyncGetChangeProofRequest,
+		responseBytes []byte,
+		onFinish func(maybe.Maybe[[]byte]),
+	) error
+	RegisterErrorHandler(handler func(error))
 }
